% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bino_logistic_cpp}
\alias{bino_logistic_cpp}
\title{Binomial Logistic Regression using Gradient Descent}
\usage{
bino_logistic_cpp(X, y, alpha = 0.01, max_iter = 1000L, tol = 1e-06)
}
\arguments{
\item{X}{A numeric matrix of predictors.}

\item{y}{A numeric vector of binary outcomes.}

\item{alpha}{A numeric value for the learning rate.}

\item{max_iter}{Maximum number of iterations for gradient descent.}

\item{tol}{Convergence tolerance.}
}
\value{
A numeric vector of estimated coefficients.
}
\description{
This function fits a logistic regression model using gradient descent.
}
