% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BL22204009-package.R
\docType{package}
\name{BL22204009-package}
\alias{BL22204009}
\alias{BL22204009-package}
\title{BL22204009: Statistical computing with R}
\description{
Several R and Rcpp function are included in this package for illustration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mekuanintsimeneh/BL22204009}
  \item Report bugs at \url{https://github.com/mekuanintsimeneh/BL22204009/issues}
}

}
\author{
\strong{Maintainer}: Mekuanint Simeneh \email{mekuanintsimeneh@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
