% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_exponential_cpp}
\alias{gibbs_exponential_cpp}
\title{Gibbs Sampling for Exponential Distribution (Rcpp)
@importFrom MASS glm
@importFrom boot boot
@importFrom DAAG cv.lm
@importFrom bootstrap bootstrap
@importFrom coda effectiveSize gelman.diag}
\usage{
gibbs_exponential_cpp(y, alpha, beta, n_iter)
}
\arguments{
\item{y}{Numeric vector of observed data.}

\item{alpha}{Prior shape parameter for the Gamma distribution.}

\item{beta}{Prior rate parameter for the Gamma distribution.}

\item{n_iter}{Number of Gibbs sampling iterations.}
}
\value{
A vector of sampled lambda values or the posterior mean if return_mean = TRUE.
}
\description{
Gibbs Sampling for Exponential Distribution (Rcpp)
@importFrom MASS glm
@importFrom boot boot
@importFrom DAAG cv.lm
@importFrom bootstrap bootstrap
@importFrom coda effectiveSize gelman.diag
}
