% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{gradient_R}
\alias{gradient_R}
\title{Gradient of the log-likelihood for Binomial Logistic Regression}
\usage{
gradient_R(X, y, beta)
}
\arguments{
\item{X}{A numeric matrix of predictors.}

\item{y}{A numeric vector of binary outcomes (0 or 1).}

\item{beta}{A numeric vector of regression coefficients.}
}
\value{
A numeric vector of gradients for the coefficients.
}
\description{
This function computes the gradient of the log-likelihood for logistic regression.
}
