% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{logistic_R}
\alias{logistic_R}
\title{Logistic function}
\usage{
logistic_R(x)
}
\arguments{
\item{x}{A numeric value or vector.}
}
\value{
Logistic transformation of input x.

Logistic transformation of input x.
}
\description{
This is a logistic sigmoid function that maps input values to probabilities between 0 and 1.
}
