% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{My Dataset}
\format{
A data frame with X rows and Y variables:
\describe{
\item{Age}{Age of the individual (numeric).}
\item{BMI}{Body Mass Index (numeric).}
\item{BloodPressure}{Blood pressure measurement (numeric).}
\item{DiabetesPedigreeFunction}{Diabetes pedigree function score (numeric).}
\item{Outcome}{Diagnosis of diabetes (1: Positive, 0: Negative).}
}
}
\source{
Source of the dataset, if applicable.
}
\usage{
mydata
}
\description{
A brief description of the dataset.
}
\keyword{datasets}
